+++ src/py21cmfast/src/interp_tables.h
@@ -22,9 +22,9 @@ double EvaluateFcoll_delta(double delta, double growthf, double sigma_min, doubl
 void init_FcollTable(double zmin, double zmax, bool x_ray);
 double EvaluatedFcolldz(double delta, double redshift, double sigma_min, double sigma_max);
 
-void initialise_Nion_Conditional_spline(float z, float Mcrit_atom, float min_density, float max_density,
-                                     float Mmin, float Mmax, float Mcond, float log10Mturn_min, float log10Mturn_max,
-                                     float log10Mturn_min_MINI, float log10Mturn_max_MINI, float Alpha_star,
+void initialise_Nion_Conditional_spline(double z, double Mcrit_atom, double min_density, double max_density,
+                                     double Mmin, double Mmax, double Mcond, double log10Mturn_min, double log10Mturn_max,
+                                     double log10Mturn_min_MINI, double log10Mturn_max_MINI, float Alpha_star,
                                      float Alpha_star_mini, float Alpha_esc, float Fstar10, float Fesc10,
                                      float Mlim_Fstar, float Mlim_Fesc, float Fstar7_MINI, float Fesc7_MINI,
                                      float Mlim_Fstar_MINI, float Mlim_Fesc_MINI, int method, int method_mini,
